migrate file_age and add some debugging stuff
