+++ procbuild/builder.py
@@ -10,25 +10,11 @@ import time
 import random
 
 from os.path import join as joinp
-from datetime import datetime, timedelta
 from glob import glob
 
 excluded = ['vanderwalt', '00_vanderwalt', 'jane_doe', 'bibderwalt', '00_intro']
 
 base_path = os.path.abspath(os.path.dirname(__file__))
-
-
-def file_age(fn):
-    """Return the age of file `fn` in minutes.  Return None is the file does
-    not exist.
-    """
-    if not os.path.exists(fn):
-        return None
-
-    modified = datetime.fromtimestamp(os.path.getmtime(fn))
-    delta = datetime.now() - modified
-
-    return delta.seconds / 60
 
 
 def cache(path='../cache'):
